Move sanitizeGraphQLName to Subgraph so it can be easily reused for subgraph extraction
